feat(#9): Implemented State class
Implement the abstract class State, which serves as a blueprint for managing game states. This class includes methods for handling user input, updating state, rendering, and disposing. I added cam, idk if we are going to use this. This class is intended to be extended by specific game states within the application.