refactor(#9): Commented out function in StartScreenState
Since im implementing main menu i dont think we are going to use StartScreenState class. I removed the previous function for adding game state and added a new function for this in GameStateManager so there is an error in the StartScreenState. This is why i commented this function out to prevent error.