feat(#9): Implemented MainMenuState class
Implemented the MenuState class, responsible for managing the menu screen in the game. When adding a new button you only need to write addButton and the text displayed inside the button. This becuase it is easier to change display of buttons. The MenuButton class handles the rest. The buttons is added to a list for easier handling. The addButton function also takes in the Texture rather than implementing this in the MenuButton class so that we can use the MenuButton class to other states if needed.