refactor(#9): Changed name and location of PlayerState
I change the name to PlayState since this better reflects the state. I also changes the location to states. I implemented GameStateManger and extended the class to state. I also just imeplemented a simple image of player so its easier to see if adding the state works since it previouslt just was black.