feat(#19): pushed optionsState in MenuState
I added a push(optionsState) so that when we click the options button we get sent to the new state. Earlier we had a problem with using push and pop becuase the screen went black when returing to menuState but i figured out we just had to remove the dispose() method right after we push, becuase it removes the state.